Add protocol definitions for Fetch API v9
This commit only adds the minimal "support" for the Fetch API v9 on the protocol level, but
does not implement the additional logic described by KIP-320 to better handle unclear leader
elections.

In particular KIP-320 allows a client to "opt-out" of the complex leader epoch validation,
and at that point only changes the Fetch API to use two different error codes as replacement
for the previous code indicating "metadata is stale".