Allow optional sharing of a single broker pool across consumers or producers
Essentially this is a refactoring in which the connection-related arguments provided
to the `consumer`/`producer` functions aren't processed by the cluster itself, but
rather the client creates the connection builder and BrokerPool instance and passes
that through to the cluster.

This allows the caller to create a suitable broker pool beforehand, and pass that instead.

If a broker pool is passed no other connection-related options must be provided, otherwise
the call will fail immediately. This is intentionally a very strict test, to avoid confusion
about the behavior. Only advanced cases would provide the broker pool, and these should also
be able to ensure no misleading other options are provided.

Note that this should be 100% API compatible up to the additional `brokerPool` function
and options.

See https://github.com/tulios/kafkajs/issues/961 for a discussion