Allow optional sharing of a single "connection pool" across consumers or producers
Essentially this is a refactoring in which the connection-related arguments provided
to the `consumer`/`producer` functions aren't processed by the cluster itself, but
rather the client creates a connection pool consisting of the connection builder and
the BrokerPool instance and passes that through to the cluster.

If a connection pool is passed no other connection-related options must be provided, otherwise
the call will fail immediately. This is intentionally a very strict test, to avoid confusion
about the behavior.

Note that this should be 100% API compatible up to the additional `connectionPool` function
and options.

The bulk of the changes needed is to ensure that the broker pool can handle being shared, which
touches a few parts:
* [TODO]: Connecting and disconnecting done through the Cluster may not lead to an actual
  change in the connection status in the broker, as that one may already be connected, or may
  still have other clusters that expect it to be connected
* [TODO]: Instrumentation events related to networking and requests need to be routed to all
  affected receivers, without leaking events intended for different receivers
* The refreshing of metadata from a broker needs to be aware that it can be triggered from
  multiple clusters at the same time, and that each of these clusters has a different set
  of target topics that need to be included.
  The `refreshMetadata` function was rewritten to provide that without introducing a huge
  lock similar to the management of the target topics in the Cluster.

See https://github.com/tulios/kafkajs/issues/961 for a discussion